requirements.txt
asgiref==3.7.2
Django==4.0.8
django-tastypie==0.14.5
Pillow==10.0.0
python-dateutil==2.8.2
python-mimeparse==1.6.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
